Перенес модели в отдельный файл
