+++ article.py
@@ -1,0 +1,20 @@
+from flask_sqlalchemy import SQLAlchemy
+from datetime import datetime
+
+db = SQLAlchemy()
+
+
+class Article(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    title = db.Column(db.String(100), nullable=False)
+    intro = db.Column(db.String(300), nullable=False)
+    text = db.Column(db.Text, nullable=False)
+    date = db.Column(db.DateTime, default=datetime.utcnow())
+
+    def __init__(self, title, intro, text):
+        self.title = title
+        self.intro = intro
+        self.text = text
+
+    def __repr__(self):
+        return '<Article %r>' % self.id
